requirements.txt
argparse
openai
python-dotenv
openai-agents
tqdm